requirements.txt
# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Gemini AI
google-genai>=1.33.0

# HTML 파싱
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 이미지 캡처
playwright>=1.40.0

# Excel 처리
pandas>=2.0.0
openpyxl>=3.1.0

python-dotenv>=1.0.0

# 기타 유틸리티
pathlib2>=2.3.0
